Preserving Families and Destruction
This is an update to the original code that will accurately model a
layer’s descendants when storing it as a class.

Originally, when passing a layer into the createClass function, all of
that layer’s descendants would become first-level sublayers of the
passed layer. This ‘flattened’ layer families where a layer’s sublayers
had sublayers of their own — and, because layers are positioned
relatively within their parent’s frames, this change would disrupt
positioning for the subLayers effected.

The new code will preserve that family tree, ensuring that node
relationships remain in place and that grandchildren and so on are
positioned correctly.

I also added a very simple kill switch for destroying the source layer
after converting it to a class.